requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
playwright==1.39.0
firecrawl-py==2.8.0
langchain==0.0.334
python-dotenv==1.0.0
loguru==0.7.2
python-multipart==0.0.6
